3.4-django-testing/django_testing/requirements.txt
django~=4.1.3
djangorestframework~=3.14.0
django-filter~=22.1
psycopg2-binary
pytest~=7.2.0
model_bakery~=1.9.0